% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_schedule.R
\name{get_conf_schedule}
\alias{get_conf_schedule}
\title{Or you can return a conference worth of scheduling
Return NCAA schedule dataframe object}
\usage{
get_conf_schedule(conf_name, yr)
}
\value{
get_conf_schedule data.frame: 'url', 'date', 'home', 'away', 'location'
}
\description{
Or you can return a conference worth of scheduling
Return NCAA schedule dataframe object
}
\examples{
\dontrun{
  get_conf_schedule('Big Ten', 2023)
  }

}
