% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beach.R
\name{get_fivb_finishes}
\alias{get_fivb_finishes}
\title{Matching Hudl's team XML files to final finishes by event}
\usage{
get_fivb_finishes()
}
\value{
get_team_schedule data.frame: 'url', 'date', 'home', 'away', 'location'
}
\description{
Matching Hudl's team XML files to final finishes by event
}
\examples{
\dontrun{
  get_team_schedule('Yale', 813, 2019)
  }

}
