% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_teams.R
\name{get_ncaaw_teams}
\alias{get_ncaaw_teams}
\title{Return NCAA teams dataframe object}
\usage{
get_ncaaw_teams()
}
\value{
get_ncaaw_teams data.frame: 'Name', 'Conference', 'ID'
}
\description{
Return NCAA teams dataframe object
}
\examples{
\dontrun{
  get_ncaaw_teams
  }

}
