% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_schedule.R
\name{get_team_schedule}
\alias{get_team_schedule}
\title{After running `get_ncaaw_teams()` you can then take the 'Name' and 'ID' to return that teams specific schedule
Return NCAA schedule dataframe object}
\usage{
get_team_schedule(team_name, team_id, yr)
}
\value{
get_team_schedule data.frame: 'url', 'date', 'home', 'away', 'location'
}
\description{
After running `get_ncaaw_teams()` you can then take the 'Name' and 'ID' to return that teams specific schedule
Return NCAA schedule dataframe object
}
\examples{
\dontrun{
  get_team_schedule(Yale, 813, 2019)
  }

}
