% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_app.R
\name{rec_app}
\alias{rec_app}
\title{Return a reception dashboard with your data}
\usage{
rec_app(data)
}
\arguments{
\item{data:}{dataset}
}
\value{
data read from datavolleyXtra::dv_readXtra()
}
\description{
Simple reception dashboard meant for indoor volleyball breaking down the reception of teams by rotation with the ability to filter by opponents. Best used if datavolleyXtra::dv_readXtra() is ran
}
\examples{
\dontrun{
  df <- datavolleyXtra::dv_readXtra("DVWs")
  rec_app(df)
  }
}
\author{
Tyler Widdison
}
