% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_end_zone_table.R
\name{rec_end_zone}
\alias{rec_end_zone}
\title{Return a reception table by end zone}
\usage{
rec_end_zone(data, team_name)
}
\arguments{
\item{data:}{dataset}

\item{team_name:}{teamname desired in dataset. Can use short names, but will return other teams with smiliar short names IE 'Texas' will return multiple teams if more than one 'Texas' team exists}
}
\value{
team_name can be string partial. data table with simple reception statistics by rotation. Returns reception attempts > 3
}
\description{
Return a reception table by end zone
}
\examples{
\dontrun{
  dv_readXtra(df, 'Rutgers')
  }
}
\author{
Tyler Widdison
}
