% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_table.R
\name{rec_table}
\alias{rec_table}
\title{Return a simple reception table}
\usage{
rec_table(data, team_name, rotation)
}
\arguments{
\item{data:}{dataset}

\item{team_name:}{teamname desired in dataset. Can use short names, but will return other teams with smiliar short names IE 'Texas' will return multiple teams}

\item{rotation:}{desired rotation}
}
\value{
data table with simple reception statistics
}
\description{
Return a simple reception table
}
\examples{
\dontrun{
  rec_table(data, 'Yale', 3)
  }

}
\author{
Tyler Widdison
}
