% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_readxtras.R
\name{dv_readXtra}
\alias{dv_readXtra}
\title{Return datavolley plays object with added columns}
\usage{
dv_readXtra(d = getwd())
}
\arguments{
\item{:}{dvw directory}
}
\value{
data.frame: the plays component of a datavolley object, as returned by \code{dv_read()} In addition to plays, it also returns the following: filename, date, receiving_team, receiving_team_id, set_won_by, home_sets_won, visiting_sets_won, match_won_by, set_won_by_id, team_won_set, match_won_by_id, team_won_match, home_setter_id, visting_setter_id, setter_id, setter_position, setter_front_back, opponent, sets_won, match_won, position, reception_quality, set_player
}
\description{
Return datavolley plays object with added columns
}
\examples{
\dontrun{
  dv_readXtra("dvws")
  }
}
\references{
\url{https://github.com/openvolley/datavolley}
}
