% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_table.R
\name{rec_table}
\alias{rec_table}
\title{Return a simple reception table}
\usage{
rec_table(data, team_name)
}
\arguments{
\item{data:}{dataset}

\item{team_name:}{teamname desired in dataset. Can use short names, but will return other teams with similar short names IE 'Texas' will return multiple teams}
}
\value{
team_name can be string partial. data table with simple reception statistics by rotation. Returns reception attempts > 3
}
\description{
Return a simple reception table
}
\examples{
\dontrun{
  rec_table(df, 'Rutgers')
  }
}
\author{
Tyler Widdison
}
